requirements.txt
Flask==1.1.2
connexion==2.7.0
dnspython==1.16.0
Flask-Cors==3.0.8
flask-mongoengine==0.9.5
pytest-flask==1.0.0
pytest==5.4.1
Werkzeug==1.0.1
python-dotenv==0.13.0
swagger-ui-bundle==0.0.6
pymongo==3.9.0